cog-agent-git/requirements.txt


numpy>=1.23
scipy>=1.9
scikit-learn>=1.2

torch>=2.0
gymnasium>=0.28
stable-baselines3>=2.0

sentence-transformers>=2.2
networkx>=3.0
matplotlib>=3.7

faiss-cpu>=1.7.4; platform_system != "Windows"
PyYAML>=6.0
